+++ src/screens/node-settings/node-settings.tsx
@@ -7,7 +7,7 @@ import { HeaderTitle } from 'src/components/header/header-title/header-title';
 import { useNavigationSetOptions } from 'src/components/header/use-navigation-set-options.hook';
 import { IconNameEnum } from 'src/components/icon/icon-name.enum';
 import { ScreenContainer } from 'src/components/screen-container/screen-container';
-import { StyledRadioGroup } from 'src/components/styled-radio-buttons-group/styled-radio-buttons-group';
+import { StyledRadioGroup } from 'src/components/styled-radio-group';
 import { ModalsEnum } from 'src/navigator/enums/modals.enum';
 import { ScreensEnum } from 'src/navigator/enums/screens.enum';
 import { useNavigation } from 'src/navigator/hooks/use-navigation.hook';
